Update colophon.xhtml
Added the year of completion for the cover artwork. 

- Source:
https://standardebooks.org/artworks/francisco-de-goya-y-lucientes/bullfight-in-a-divided-ring

Note:
Francisco Goya died in 1828. So, the 1867 date seems wrong. But that's the date I found on the website (Standard Ebooks). On Google Arts and Culture, the date is 1800/1829. I'll leave it up to you to choose whichever of the two seems best.

Google Arts and Culture link: https://artsandculture.google.com/asset/bullfight-in-a-divided-ring-goya-francisco-de-goya-y-lucientes/kAFXj5ds2VU6zw